requirements.txt
biopython==1.79
tqdm==4.62.3